Task Management

+++ Modules/TaskManagement/Resources/views/admin/activity/index.blade.php
@@ -61,22 +61,34 @@
                                         </ul>
                                     </td>
                                     <td>
-                                        <a data-bs-type="edit"
-                                           href="{{route('admin.taskManagement.activity.edit',$activity)}}"
-                                           title="सम्पादन गर्नुहोस्"
-                                           class="btn btn-xs btn-outline-primary {{get_setting('Pin')?'confirm_pin':''}}">
-                                            <i class="fa fa-edit"></i>
-                                        </a>
-                                        <form action="{{route('admin.taskManagement.activity.destroy',$activity)}}"
-                                              method="post">
-                                            @csrf
-                                            @method('delete')
-                                            <button data-bs-type="delete"
-                                                    class="btn btn-xs btn-outline-danger {{get_setting('Pin')?'confirm_pin':'show_confirm'}}"
-                                                    title="मेटाउनु होस्">
-                                                <i class="fa fa-trash"></i>
-                                            </button>
-                                        </form>
+                                        @can('taskActivity_access')
+                                            <a
+                                                href="{{route('admin.taskManagement.activity.show',$activity)}}"
+                                                title="हेर्नुहोस"
+                                                class="btn btn-xs btn-outline-success">
+                                                <i class="fa fa-eye"></i>
+                                            </a>
+                                        @endcan
+                                        @can('taskActivity_edit')
+                                            <a
+                                                href="{{route('admin.taskManagement.activity.edit',$activity)}}"
+                                                title="सम्पादन गर्नुहोस्"
+                                                class="btn btn-xs btn-outline-primary {{get_setting('Pin')?'confirm_pin':''}}">
+                                                <i class="fa fa-edit"></i>
+                                            </a>
+                                        @endcan
+                                        @can('taskActivity_delete')
+                                            <form action="{{route('admin.taskManagement.activity.destroy',$activity)}}"
+                                                  method="post">
+                                                @csrf
+                                                @method('delete')
+                                                <button data-bs-type="delete"
+                                                        class="btn btn-xs btn-outline-danger {{get_setting('Pin')?'confirm_pin':'show_confirm'}}"
+                                                        title="मेटाउनु होस्">
+                                                    <i class="fa fa-trash"></i>
+                                                </button>
+                                            </form>
+                                        @endcan
                                     </td>
                                 </tr>
                             @empty
